+++ module/rules.php
@@ -79,7 +79,7 @@ print '</select>';
 print dol_get_fiche_end();
 
 print dolicurateConfigBlock('rules');
-print '<script src="'.dol_buildpath('/dolicurate/js/dolicurate-rules.js', 1).'?v='.urlencode(getDolGlobalString('MAIN_MODULE_DOLICURATE_VERSION', '1.0.0')).'"></script>';
+print '<script src="'.dol_buildpath('/dolicurate/js/dolicurate-rules.js', 1).'?v='.urlencode(dolicurateAssetVersion('/dolicurate/js/dolicurate-rules.js')).'"></script>';
 
 llxFooter();
 $db->close();